+++ packages/article-summary/src/article-summary-content.js
@@ -2,7 +2,7 @@ import React from "react";
 import { Text } from "react-native";
 import PropTypes from "prop-types";
 import { propTypes as treePropType } from "@times-components-native/markup-forest";
-import { renderAst } from "./article-summary";
+import { renderAst as defaultRenderAst } from "./article-summary";
 import styles from "./styles";
 
 const ArticleSummaryContent = ({
@@ -11,6 +11,7 @@ const ArticleSummaryContent = ({
   style,
   whiteSpaceHeight,
   initialLines = 2,
+  renderAst = defaultRenderAst,
 }) => {
   const lineHeight = (style && style.lineHeight) || styles.text.lineHeight;
   const numberOfLinesToRender =
