+++ packages/edition-slices/src/tiles/shared/index.js
@@ -2,7 +2,6 @@ import { getTileImage, getTileSummary, getTileStrapline } from "./utils";
 import TileLink from "./tile-link";
 import TileStar from "./tile-star";
 import TileSummary from "./tile-summary";
-import FrontTileSummary from "./front-tile-summary";
 import TileImage from "./tile-image";
 import withTileTracking from "./tile-tracking-events";
 import withPuzzleTileTracking from "./puzzle-tracking-events";
@@ -14,7 +13,6 @@ export {
   TileLink,
   TileStar,
   TileSummary,
-  FrontTileSummary,
   TileImage,
   withPuzzleTileTracking,
   withTileTracking,
