+++ packages/edition-slices/src/tiles/tile-al-front/index.js
@@ -2,11 +2,11 @@
 import React from "react";
 import PropTypes from "prop-types";
 import { editionBreakpoints } from "@times-components-native/styleguide";
+import { FrontTileSummary } from "@times-components-native/front-page";
 import {
   getTileImage,
   getTileSummary,
   TileLink,
-  FrontTileSummary,
   TileImage,
   withTileTracking,
 } from "../shared";
