+++ packages/front-page/index.ts
@@ -1,2 +1,1 @@
-export { default as FrontArticleSummary } from "./front-article-summary";
-export { default as FrontArticleSummaryContent } from "./front-article-summary-content";
+export { default as FrontTileSummary } from "./front-tile-summary";
